rename reducer and actions
